B18 Decompile (0.18.6527.35959)

+++ RimWorld/IncidentCategoryEntry.cs
@@ -12,7 +12,7 @@ namespace RimWorld
 
 		public void LoadDataFromXmlCustom(XmlNode xmlRoot)
 		{
-			this.category = (IncidentCategory)((byte)ParseHelper.FromString(xmlRoot.Name, typeof(IncidentCategory)));
+			this.category = (IncidentCategory)ParseHelper.FromString(xmlRoot.Name, typeof(IncidentCategory));
 			this.weight = (float)ParseHelper.FromString(xmlRoot.FirstChild.Value, typeof(float));
 		}
 	}
